+++ openid/src/main/java/fish/payara/security/openid/domain/OpenIdContextImpl.java
@@ -1,7 +1,7 @@
 /*
  *  DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
  * 
- *  Copyright (c) [2018] Payara Foundation and/or its affiliates. All rights reserved.
+ *  Copyright (c) [2018-2020] Payara Foundation and/or its affiliates. All rights reserved.
  * 
  *  The contents of this file are subject to the terms of either the GNU
  *  General Public License Version 2 only ("GPL") or the Common Development
@@ -65,7 +65,7 @@ public class OpenIdContextImpl implements OpenIdContext {
     private AccessToken accessToken;
     private IdentityToken identityToken;
     private RefreshToken refreshToken;
-    private Integer expiresIn;
+    private Long expiresIn;
     private JsonObject claims;
     private JsonObject providerMetadata;
 
@@ -129,11 +129,11 @@ public class OpenIdContextImpl implements OpenIdContext {
     }
 
     @Override
-    public Optional<Integer> getExpiresIn() {
+    public Optional<Long> getExpiresIn() {
         return Optional.ofNullable(expiresIn);
     }
 
-    public void setExpiresIn(Integer expiresIn) {
+    public void setExpiresIn(Long expiresIn) {
         this.expiresIn = expiresIn;
     }
